dpaa2: rc: extend ICID to 32-bits for VFIO support
While the ICID is only documented as a 16-bit value, 32-bits are
actually used (taking over the 'reserved' bits next to it).
The top 16 bits are then used to differentiate containers running
under a VM.

+++ sys/dev/dpaa2/dpaa2_mcp.h
@@ -355,7 +355,7 @@ struct dpaa2_rc_attr {
 	uint32_t	cont_id;
 	uint32_t	portal_id;
 	uint32_t	options;
-	uint16_t	icid;
+	uint32_t	icid;
 };
 
 /**
